requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0

# LangChain and LangGraph
langgraph==1.0.5
langchain-google-genai==4.0.0

# Gemini LLM
google-generativeai==0.3.2

# Embeddings
sentence-transformers==2.3.1
torch

# Vector Store
faiss-cpu==1.13.1

# BM25
rank-bm25==0.2.2

# Document Processing
PyMuPDF==1.23.21  # PDF processing
python-docx==1.1.0  # DOCX processing

# Utilities
pydantic==2.12.5
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyYAML==6.0.1

# HTTP Client for Frontend
requests==2.31.0
websockets==12.0
